// packages/cdk/src/functions/types.d.ts
export declare enum FunctionRuntime {
    NODE = "node",
    PYTHON = "python",
    DOTNET = "dotnet",
    JAVA = "java",
    POWERSHELL = "powershell",
    CUSTOM = "custom"
}